+++ spec/cucumber/core/compiler_spec.rb
@@ -137,19 +137,11 @@ module Cucumber::Core
 
         compile(gherkin_documents) do |visitor|
           expect( visitor ).to receive(:test_step) do |test_step|
-            visit_source(test_step) do |source_visitor|
-              expect( source_visitor ).to receive(:step) do |step|
-                expect(step.text).to eq 'passing 1 with 2'
-              end
-            end
+            expect(test_step.text).to eq 'passing 1 with 2'
           end.once.ordered
 
           expect( visitor ).to receive(:test_step) do |test_step|
-            visit_source(test_step) do |source_visitor|
-              expect( source_visitor ).to receive(:step) do |step|
-                expect(step.text).to eq 'as well as 3'
-              end
-            end
+            expect(test_step.text).to eq 'as well as 3'
           end.once.ordered
 
           expect( visitor ).to receive(:done).once.ordered
@@ -158,7 +150,7 @@ module Cucumber::Core
     end
 
     context 'empty scenarios' do
-      it 'does not create test cases for them' do
+      it 'does create test cases for them' do
         gherkin_documents = [
           gherkin do
             feature do
@@ -168,90 +160,10 @@ module Cucumber::Core
           end
         ]
         compile(gherkin_documents) do |visitor|
-          expect( visitor ).to receive(:test_case).never
+          expect( visitor ).to receive(:test_case).once.ordered
           expect( visitor ).to receive(:done).once.ordered
         end
       end
-    end
-
-    describe Compiler::FeatureCompiler do
-      let(:receiver) { double('receiver') }
-      let(:compiler) { Compiler::FeatureCompiler.new(receiver) }
-
-      context "a scenario with a background" do
-        stubs(:feature,
-                :background,
-                  :background_step,
-                :scenario,
-                  :scenario_step)
-
-        it "sets the source correctly on the test steps" do
-          expect( receiver ).to receive(:on_background_step).with(
-            [feature, background, background_step]
-          )
-          expect( receiver ).to receive(:on_step).with(
-            [feature, scenario, scenario_step]
-          )
-          expect( receiver ).to receive(:on_test_case).with(
-            [feature, scenario]
-          )
-          compiler.feature(feature) do |f|
-            f.background(background) do |b|
-              b.step background_step
-            end
-            f.scenario(scenario) do |s|
-              s.step scenario_step
-            end
-          end
-        end
-      end
-
-      context "a scenario outline" do
-        stubs(:feature,
-                :background,
-                  :background_step,
-                :scenario_outline,
-                  :outline_step,
-                  :examples_table_1,
-                    :examples_table_1_row_1,
-                      :outline_ast_step,
-                  :examples_table_2,
-                    :examples_table_2_row_1,
-             )
-
-        it "sets the source correctly on the test steps" do
-          allow( outline_step ).to receive(:to_step) { outline_ast_step }
-          expect( receiver ).to receive(:on_step).with(
-            [feature, scenario_outline, examples_table_1, examples_table_1_row_1, outline_ast_step]
-          ).ordered
-          expect( receiver ).to receive(:on_test_case).with(
-            [feature, scenario_outline, examples_table_1, examples_table_1_row_1]
-          ).ordered
-          expect( receiver ).to receive(:on_step).with(
-            [feature, scenario_outline, examples_table_2, examples_table_2_row_1, outline_ast_step]
-          ).ordered
-          expect( receiver ).to receive(:on_test_case).with(
-            [feature, scenario_outline, examples_table_2, examples_table_2_row_1]
-          ).ordered
-          compiler.feature(feature) do |f|
-            f.scenario_outline(scenario_outline) do |o|
-              o.outline_step outline_step
-              o.examples_table(examples_table_1) do |t|
-                t.examples_table_row(examples_table_1_row_1)
-              end
-              o.examples_table(examples_table_2) do |t|
-                t.examples_table_row(examples_table_2_row_1)
-              end
-            end
-          end
-        end
-      end
-    end
-
-    def visit_source(node)
-      visitor = double.as_null_object
-      yield visitor
-      node.describe_source_to(visitor)
     end
 
     def compile(gherkin_documents)
@@ -264,4 +176,3 @@ module Cucumber::Core
 
   end
 end
-